#35 ♻️ refactor: 캘린더 컴포넌트 리팩토링
