--- src/common/components/Calendar/index.ts
@@ -1,1 +1,0 @@
-export * from './Calendar';
