+++ src/features/Calendar/components/Calendnar/Calendar.tsx
@@ -1,8 +1,12 @@
-import React, { useState, useCallback, useRef } from 'react';
+import { useState, useCallback, useRef, Dispatch, SetStateAction } from 'react';
 import styled from '@emotion/styled';
-import { AnimatePresence, motion } from 'framer-motion';
+import CalendarHeader from './CalendarHeader';
 
-const Calendar = () => {
+interface Props {
+  setIsRecordListShown: Dispatch<SetStateAction<boolean>>;
+}
+
+const Calendar = ({ setIsRecordListShown }: Props) => {
   const today = {
     year: new Date().getFullYear(),
     month: new Date().getMonth() + 1,
@@ -19,7 +23,6 @@ const Calendar = () => {
 
   const dateNumListFrameRef = useRef<null | HTMLDivElement>(null);
   const dateNumListRef = useRef<null | HTMLDivElement>(null);
-  const [isRecordListShown, setIsRecordListShown] = useState(false);
 
   /**
    * 숫자 n을 입력받아 1부터 n까지의 array를 생성해 주는 함수
@@ -29,28 +32,6 @@ const Calendar = () => {
   const createArray1ToN = (n: number) => {
     return Array.from({ length: n }, (_, index) => index + 1);
   };
-
-  /**
-   * 이전 달 보기 함수
-   */
-  const prevMonth = useCallback(() => {
-    setSelectedDate((prev) => ({
-      year: prev.month === 1 ? prev.year - 1 : prev.year,
-      month: prev.month === 1 ? 12 : prev.month - 1,
-      date: 1,
-    }));
-  }, []);
-
-  /**
-   * 다음 달 보기 함수
-   */
-  const nextMonth = useCallback(() => {
-    setSelectedDate((prev) => ({
-      year: prev.month === 12 ? prev.year + 1 : prev.year,
-      month: prev.month === 12 ? 1 : prev.month + 1,
-      date: 1,
-    }));
-  }, []);
 
   /**
    * 달력의 날짜 클릭 시 해당 날짜로 포커스 시키는 함수
@@ -74,53 +55,13 @@ const Calendar = () => {
         dateNumListRef.current.style.transform = `translateY(${-upValue}px)`;
       }
     },
-    [firstDayOfMonth],
+    [firstDayOfMonth, setIsRecordListShown],
   );
-
-  // // drag & drop
-  // const [touchY, setTouchY] = useState(0);
-
-  // const onTouchStart = (e: React.TouchEvent<HTMLDivElement>) => {
-  //   setTouchY(e.changedTouches[0].pageY);
-  // };
-
-  // const onTouchEnd = (e: React.TouchEvent<HTMLDivElement>) => {
-  //   const distanceY = touchY - e.changedTouches[0].pageY;
-  //   console.log(distanceY);
-  //   // const targetUl = e.target.closest('ul');
-  //   // const targetUlWidth = targetUl.offsetWidth / 2;
-  //   // const newLeft = Math.abs(parseFloat(targetUl.style.left)) + distanceY;
-
-  //   // if (newLeft < 0) {
-  //   //   targetUl.style.left = '0px';
-  //   // } else if (newLeft < targetUlWidth) {
-  //   //   targetUl.style.left = `-${newLeft}px`;
-  //   // } else {
-  //   //   targetUl.style.left = `-${targetUlWidth}px`;
-  //   // }
-  // };
 
   return (
     <StyledCalendar>
       {/* 달력 헤더 영역(달, 년도, 화살표 아이콘) Start */}
-      <StyledCalendarHeader>
-        <StyledMonthYear>
-          <StyledMonth>
-            {selectedDate.month < 10 && '0'}
-            {selectedDate.month}
-          </StyledMonth>
-          <StyledYear>{selectedDate.year}</StyledYear>
-        </StyledMonthYear>
-
-        <div>
-          <button type='button' onClick={() => prevMonth()}>
-            이전
-          </button>
-          <button type='button' onClick={() => nextMonth()}>
-            다음
-          </button>
-        </div>
-      </StyledCalendarHeader>
+      <CalendarHeader selectedDate={selectedDate} setSelectedDate={setSelectedDate} />
       {/* 달력 헤더 영역(달, 년도, 화살표 아이콘) End */}
 
       {/* 요일 영역(SUN, MON, ... , SAT) Start */}
@@ -158,29 +99,6 @@ const Calendar = () => {
         </StyledDateNumList>
       </StyledDateNumListFrame>
       {/* 날짜 영역 End */}
-
-      <AnimatePresence>
-        {isRecordListShown && (
-          <StyledRecordList
-            variants={recordListVar}
-            initial='invisible'
-            animate='visible'
-            exit='exit'
-            // onTouchStart={(e) => onTouchStart(e)}
-            // onTouchEnd={(e) => onTouchEnd(e)}
-          >
-            {[0, 1, 2].map((record: number) => (
-              <StyledRecord key={record}>
-                <StyledRecordImage />
-                <div>
-                  <StyledRecordTitle>WONDERWALL</StyledRecordTitle>
-                  <StyledRecordArtist>OASIS</StyledRecordArtist>
-                </div>
-              </StyledRecord>
-            ))}
-          </StyledRecordList>
-        )}
-      </AnimatePresence>
     </StyledCalendar>
   );
 };
@@ -188,30 +106,7 @@ const Calendar = () => {
 const StyledCalendar = styled.div`
   width: 100%;
   color: ${(props) => props.theme.color.white};
-  padding-bottom: 15rem;
-`;
-
-const StyledCalendarHeader = styled.div`
-  display: flex;
-  justify-content: space-between;
-  align-items: flex-end;
-  margin-top: 2rem;
-`;
-
-const StyledMonthYear = styled.div`
-  display: flex;
-  align-items: flex-end;
-  font-weight: 600;
-  gap: 0.6rem;
-`;
-
-const StyledMonth = styled.div`
-  font-size: 8rem;
-  line-height: 6.8rem;
-`;
-
-const StyledYear = styled.div`
-  font-size: 1.8rem;
+  /* padding-bottom: 15rem; */
 `;
 
 const StyledDayList = styled.div`
@@ -254,47 +149,5 @@ const StyledDateNum = styled.div<{ isSelected: boolean; isToday: boolean }>`
   transition: border 0.1s ease-in-out;
   cursor: pointer;
 `;
-
-const StyledRecordList = styled(motion.div)`
-  width: 100%;
-  /* height: 100rem; */
-  margin-top: 2rem;
-  /* background-color: ${(props) => props.theme.color.gray11}; */
-`;
-
-const StyledRecord = styled.div`
-  display: flex;
-  align-items: center;
-  border-top: 1px solid ${(props) => props.theme.color.gray08};
-  padding: 1.2rem 1rem;
-`;
-
-const StyledRecordImage = styled.div`
-  width: 6.4rem;
-  aspect-ratio: 1/1;
-  background-color: ${(props) => props.theme.color.gray11};
-  margin-right: 1.6rem;
-`;
-
-const StyledRecordTitle = styled.div`
-  ${(props) => props.theme.font.bold16}
-`;
-
-const StyledRecordArtist = styled.div`
-  ${(props) => props.theme.font.medium14}
-  color: ${(props) => props.theme.color.gray08};
-`;
-
-const recordListVar = {
-  invisible: {
-    opacity: 0,
-  },
-  visible: {
-    opacity: 1,
-  },
-  exit: {
-    opacity: 0,
-  },
-};
 
 export default Calendar;
