+++ src/features/Calendar/components/Calendnar/index.ts
@@ -1,0 +1,3 @@
+import Calendar from './Calendar';
+
+export default Calendar;
