+++ src/features/Calendar/components/RecordList/index.ts
@@ -1,0 +1,3 @@
+import RecordList from './RecordList';
+
+export default RecordList;
