+++ src/pages/index.tsx
@@ -3,24 +3,34 @@ import { CalendarPageFooter } from '@common/components/Footer';
 import { CalendarPageNavigation } from '@common/components/Navigation';
 import { SvgIcon } from '@common/components/SvgIcon';
 import styled from '@emotion/styled';
+import Calendar from '@features/Calendar/components/Calendnar';
+import RecordList from '@features/Calendar/components/RecordList';
 import { theme } from '@styles/theme';
+import { useState } from 'react';
 
 const Home = () => {
+  const [isRecordListShown, setIsRecordListShown] = useState(false);
   return (
     <>
       <CalendarPageNavigation />
-      <div>캘린더 영역</div>
+
+      <Calendar setIsRecordListShown={setIsRecordListShown} />
+
+      <RecordList isRecordListShown={isRecordListShown} />
+
       <StyledContainer>
         <div>
           <StyledDate>07</StyledDate>
-          <StyledWord>DAY</StyledWord>
+          <StyledWord style={{ marginRight: '1.2rem' }}>DAY</StyledWord>
           <StyledDate>12</StyledDate>
           <StyledWord>RECORD</StyledWord>
         </div>
+
         <StyledRecordButton>
-          <SvgIcon id='chevronDown2' />
+          <SvgIcon id='plus_black' />
         </StyledRecordButton>
       </StyledContainer>
+
       <CalendarPageFooter />
     </>
   );
@@ -29,14 +39,13 @@ const Home = () => {
 // NOTE: Footer 쪽 스타일과 상당히 유사해서 하나의 컴포넌트로 뺄 수 있을 것 같음
 const StyledContainer = styled.div`
   position: absolute;
-  bottom: 0;
-  padding-bottom: 6.8rem;
-  padding-left: 2.4rem;
-  padding-right: 2.4rem;
-  width: calc(100% - 4.8rem);
+  bottom: 8.2rem;
   display: flex;
   justify-content: space-between;
   align-items: center;
+  padding-left: 2.4rem;
+  padding-right: 2.4rem;
+  width: calc(100% - 4.8rem);
 `;
 const StyledDate = styled.span`
   ${theme.font.bold32}
